Extract command deduplication steps from the TransactionCommitter
CHANGELOG_BEGIN

CHANGELOG_END
